+++ src/extension_instance.ml
@@ -112,6 +112,12 @@ end = struct
 
   let server_options sandbox =
     let args = Settings.(get server_args_setting) |> Option.value ~default:[] in
+    let args =
+      match Settings.get Settings.dune_context_setting with
+      | None -> args
+      | Some context -> "--context" :: context :: args
+    in
+
     let command = Sandbox.get_command sandbox "ocamllsp" args in
     Cmd.log command;
     let env =